Consistent Publisher handling from Verifier
